requirements.txt
django
djangorestframework
django-two-factor-auth
psycopg2
